# Optimize resampling operations