The resampling operations we copied from gridtools are all in a massive monolithic numba-ized function, which has several disadvantages, which I'll try to outline:

1) Huge functions with deep nesting are unreadable
2) The entire function needs to be compiled in one go which slows down the initial compile step (takes about ~5 seconds)
3) The JIT compiler has to validate types and since np.sqrt is apparently only defined for float types all the data **must** be of float type even for downsampling methods which do not use ``np.sqrt``.

While splitting the huge function into individual functions should only be seen as an initial step towards refactoring this code, it already has two major advantages:

1) No super long compile step to begin with.
2) No requirement to cast the input arrays to floats unless using downsampling with standard deviation and variance methods

When running resampling on the landsat data (using mean downsampling) the timings before and after this PR are:

PRE:

  * Compilation: 6.88 s
  * Downsampling: 2.58 s

POST:

  * Compilation: 1.56 s
  * Downsampling: 1.26 s