@philippjfr 2017-10-03
Okay, I didn't stop there and enabled numba parallelization. On my machine the results are now:

* PRE: 2.58 s
* POST (not parallel): 1.26 s
* POST(parallel): 0.515 s